PacientesEliminados agregado, detalle,editar modif

+++ front/src/components/Panel Admin/DetallePaciente.jsx
@@ -18,8 +18,14 @@ import KeyIcon from "@mui/icons-material/Key";
 import QuestionMarkIcon from "@mui/icons-material/QuestionMark";
 import Grid3x3Icon from "@mui/icons-material/Grid3x3";
 const DetallePaciente = () => {
-  const { fetchPatients, patientDetail, fetchPatientByEmail, deletePatient } =
-    useContext(Context)[1];
+  const {
+    restorePatient,
+    fetchPatients,
+    patientDetail,
+    fetchPatientByEmail,
+    deletePatient,
+    fetchSoftDeletedPatient,
+  } = useContext(Context)[1];
   const { setVista, email } = useContext(Context)[6];
 
   const [loading, setLoading] = useState(true);
@@ -33,39 +39,31 @@ const DetallePaciente = () => {
   }, [patientDetail]);
 
   const handleClickDelete = (id) => {
-    deletePatient(id)
-      .then(() => {
-        // Eliminación exitosa, actualizar la lista de pacientes
-        fetchPatients();
-        alert("El paciente ha sido eliminado exitosamente.");
-      })
-      .catch((error) => {
-        console.log("Error al eliminar el paciente:", error);
-        // Manejar el error de eliminación del paciente
-      });
+    if (patientDetail.deletedAt) {
+      restorePatient(id)
+        .then(() => {
+          // Eliminación exitosa, actualizar la lista de pacientes
+          fetchSoftDeletedPatient();
+          fetchPatients();
+          alert("El paciente ha sido restaurado exitosamente.");
+        })
+        .catch((error) => {
+          console.log("Error al restaurar el paciente:", error);
+          // Manejar el error de eliminación del paciente
+        });
+    } else {
+      deletePatient(id)
+        .then(() => {
+          // Eliminación exitosa, actualizar la lista de pacientes
+          fetchPatients();
+          alert("El paciente ha sido eliminado exitosamente.");
+        })
+        .catch((error) => {
+          console.log("Error al eliminar el paciente:", error);
+          // Manejar el error de eliminación del paciente
+        });
+    }
   };
-  // const deletePatient = async (patientId) => {
-  //   try {
-  //     // Llamada para eliminar el paciente
-  //     await axios.delete(`/detail/${patientId}`);
-  //   } catch (error) {
-  //     // Manejo de errores en caso de fallo en la eliminación
-  //     console.error( error);
-  //     throw new Error("Fallo en eliminar paciente");
-  //   }
-  // };
-
-  // const handleDelete = async () => {
-  //   try {
-  //     await deletePatient(patientDetail.id);
-  //     alert("La cuenta se eliminó correctamente");
-  //     await fetchPatients(); // Llama a la función para obtener la lista actualizada de pacientes
-  //     setVista(1);
-  //   } catch (error) {
-  //     console.log("Error al eliminar el paciente:", error);
-  //     // Manejar el error al eliminar el paciente
-  //   }
-  // };
 
   return (
     <>
@@ -113,13 +111,13 @@ const DetallePaciente = () => {
         </ListItem>
 
         <ListItem>
-  <ListItemAvatar>
-    <Avatar>
-      <KeyIcon />
-    </Avatar>
-  </ListItemAvatar>
-  <ListItemText secondary="password" primary="****************" />
-</ListItem>
+          <ListItemAvatar>
+            <Avatar>
+              <KeyIcon />
+            </Avatar>
+          </ListItemAvatar>
+          <ListItemText secondary="password" primary="****************" />
+        </ListItem>
 
         <ListItem>
           <ListItemAvatar>
@@ -147,7 +145,7 @@ const DetallePaciente = () => {
       >
         <Button
           variant="outlined"
-          onClick={() => setVista(1)}
+          onClick={() => setVista(patientDetail.deletedAt === null ? 1 : 3)}
           style={{ marginRight: "10px" }}
         >
           Volver
@@ -161,7 +159,7 @@ const DetallePaciente = () => {
             setVista(1);
           }}
         >
-          Eliminar
+          {patientDetail.deletedAt ? "Restaurar" : "Eliminar"}
         </Button>
       </Box>
     </>
